Add files via upload
The standalone MAME executable produces DAT files when it's run with a commandline option, but that is not possible with the libretro MAME cores. Once an XML DAT is part of the repo we have a canonical reference but we can also can enhance the DAT, for example by backporting the 'isrunnable' XML tag from later MAME XML DATs.