Create mame2003-updates-XML.dat
This is @Riverstorm 's updates DAT file. It will need updates from the NeoGeo driver (this will happen soon) but is otherwise up to date. @Arcadez if it's OK with you to add this here I think it will help move forward with testing to have it as part of the repository.