Update libretro.c
These defaults seem to be reversed compared to the intention.

The intention can be seen here https://github.com/arcadez/mame2003-plus-libretro/blob/master/src/libretro/libretro.c#L59-L63

The intention is that low-powered systems have the speedhacks turned on by default. However the default that is sent to libretro is backwards, so that the low-powered systems **do not** get the speedhack by default but everyone else does.

My question @arcadez is -- should the default be to turn this on for everyone? Is there any reason this is the way it is now?