revert - most of the game sounds working, while we fix this up.
right now chip 1 plays most of the game sounds so I'm going to push this for now. chip2 plays explosions...... I'm going to try to disable the first chip and only hook it to the second chip to see is those sounds work without chip1. I think we just need to independent chips.