New overlay type, scaling factors, opacity control
Can now scale artwork by more than a factor of 2, though need fast hardware.
New trianle type.
New artwork option to alter the opacity of any built-in overlay.
Handling of tags, so user can, for example, set a tag for cocktail mode
on an overlay, and one for portrait, and in game code choose which to use.