Updated overlays for deadeye, gypsyjug.
For Gypsy Juggler, people often overlay coloured pants on the juggler.
The arcade game did not have these, as shown in video in the comments.
The new  overlay follows what is shown in the video.

For deadeye, the overlay is often shown with blue trousers on the player.
This actually ruins game play and is likely a misplacement of blue that
was meant to go on the text below the game field. The game field should
be all white as with Gypsy Juggler. I can find no sources for colours,
so have assumed colours as in Gypsy Juggler which does have a source
mentioned in the code.