Updated overlays for geebee, navalone, kaitei, sos
A lot of work to get Geebee arcade accurate, and also support
upright and cocktail mode. Careful examination of flyers and
existing arcade machines. The cocktail version has more colour
around the edges, but no colour in the centre. The upright
version has more colour in the top centre, but is white at
the bottom. We handle both overlays, depending on dip switch
(restart required).

For Navalone/Navarone I can find no evidence of there ever
being an overlay, though there is a common look and feel used
on many implementations of an overlay, so a similar look and
feel has been added here in that spirit. There is also an
overlay for cocktail mode that is less impressive and less
complex due to the screen reversal process limiting what an
overlay can do on such setups. We handle both overlays,
depending on dip switch (restart required).

For Kaitei I can find no evidence of there ever being an
overlay. I have created one for it as this game deserves a
good overlay as it is a gem that few people know about.
There is also an overlay for cocktail mode that is less
impressive and less complex due to the screen reversal
process limiting what an overlay can do on such setups.
We handle both overlays, depending on dip switch
(restart required).

For SOS I can find no evidence of there ever being an
overlay. I have created a simple one for it here. It works
in both upright and cocktail mode.