Update some parts of Taitosj to MAME 106
Many improvements to the emulation of the games in this driver

MAME WIP

Alpine Ski
0.96u2: Nicola Salmoria fixed sprite hardware collision detection. 
This fixes bonuses in Alpine Ski.

Bio Attack
0.96: Peter Sealy fixed collision detection in Bio Attack 
and added missing RAM to the Taito SJ memory map.

Elevator Action
0.96u2: Nicola Salmoria fixed several sprite priority issues in Elevator Action 
and the car wrapping around at the end of a level. Added sprite clipping at the sides of the screen. 
This was verified to happen on the real board, and it affects bullets wrapping around in Elevator Action. 
I'm only 99% sure that the clipping is correct this way.

Highway Race
0.96u2: Nicola Salmoria fixed several sprite priority issues in High Way Race.

Jungle Hunt / Jungle king / Pirate Pete
0.96u2: Nicola Salmoria fixed sprite hardware collision detection. This fixes the last round in Pirate Pete.

Space Cruiser
0.99u7: Pierpaolo Prazzoli fixed Space Cruiser from resetting on the asteroids level when you die.

Space Seeker
Fixed game wont coin up and start by rolling back an earlier commit which added an extra
input the game doesn't have.

The Tin Star
0.96u2: Nicola Salmoria fixed several sprite priority issues in Tin Star.
0.183: maziac and AJR added "P1/2 Continue" button to Space Cruiser. 
Note: The game has a dedicated button to continue the game if all lives are lost. 
The player has to enter a coin and press the "Continue" button

Water Ski
0.96u2: Nicola Salmoria fixed sprite hardware collision detection. 
This fixes end of level 1 in Water Ski.

Wild Western
0.103u3: Nicola Salmoria fixed the Wild Western bonus screen, 
where the horse would not scroll into view.

Kick Start - Wheelie King
0.106u11: Canim fixed P2 inputs in Kick Start. Removed 3rd coin slot.

All games
0.92: Added Coin Lockout to prevent a crash if more than 9 coins were added.