New working game Woodpecker
0.101u3: Pierpaolo Prazzoli fixed inputs in Woodpecker. Added Button1 for Player 1/2. Removed 'Difficulty' and a 'Unknown' dipswitch.

0.88u4: Tomasz Slanina and Pierpaolo Prazzoli fixed Woodpecker - Game now playable. Added new cpu1 rom ($b000), gfx1/2 roms 
and proms ($0, 20). 8th November 2004: Guru - You've probably noticed in MAME 0.88u4 that Woodpecker is now working. 
Here's a few pics of the protection module after I unearthed it. Actually 'unearthed' might give you a false impression. 
You could be mistaken for thinking that the mess on top of the PCB is sand. It's not though, it's epoxy resin. 
Though it seems like the manufacturer forgot to read the resin mixing instructions because several parts came off in BIG chunks. 
Also the resin didn't fully encase the PCB. I guess they ran out of resin or something. Lucky for us!

0.88u3: Guru and Dan Even added 'Woodpecker (set 1)' (Amenip (Palcom Queen River) 1981) and clone (set 2).